+++ petropy/__init__.py
@@ -1,5 +1,5 @@
 from .datasets import log_data
-from .download import ul_lands_download, kgs_download
+from .download import ul_lands_download, kgs_download, create_log_inventory_table
 from .electrofacies import electrofacies
 from .graphs import LogViewer
 from .log import Log