+++ petropy/graphs.py
@@ -37,13 +37,26 @@ class LogViewer(object):
         A `Log` object with curve data to display in the LogViewer
     template_xml_path : str (default None)
         Path to xml template.
-    template_defaults : str {'raw', 'multimin_oil', 'multimin_gas', 'multimin_oil_sum', 'multimin_gas_sum'} (default None)
+    template_defaults : str
         Name of default template options. Uses prebuilt template to
         display data
     top : float (default None)
         Setting to set initial top of graph
     height : float (default None)
         Setting to set amout of feet to display
+
+    Note
+    ----
+    template_defaults string must be from these options
+
+    .. code-block:: bash
+
+        raw
+        multimin_oil
+        multimin_gas
+        multimin_oil_sum
+        multimin_gas_sum
+
 
     Examples
     --------
@@ -145,39 +158,55 @@ class LogViewer(object):
 
         ### private parameters for graphically editing curves ###
 
-        self._edit_curve = None # stores display name of curve
-        self._edit_curve_lines = {} # stores matplotlib line objects by display name
-        self._edit_lock = False # stores bool to show downclick to edit curve
+        # stores display name of curve
+        self._edit_curve = None
 
-        self._radio_button = None # radio button for editing modes
+        # stores matplotlib line objects by display name
+        self._edit_curve_lines = {}
 
-        self._display_name_to_curve_name = {} # display name to log column name dictionary
+        # stores bool to show downclick to edit curve
+        self._edit_lock = False
+
+        # radio button for editing modes
+        self._radio_button = None
+
+        # display name to log column name dictionary
+        self._display_name_to_curve_name = {}
 
         default_templates_paths = {
             'raw': 'default_raw_template.xml',
             'multimin_oil': 'default_multimin_oil_template.xml',
-            'multimin_oil_sum': 'default_multimin_oil_sum_template.xml',
+            'multimin_oil_sum':'default_multimin_oil_sum_template.xml',
+            'full_oil': 'default_oil_full_template.xml',
             'multimin_gas': 'default_multimin_gas_template.xml',
+            'multimin_gas_sum': 'default_multimin_gas_sum_template.xml',
+            'full_gas': 'default_gas_full_template.xml'
         }
 
         file_dir = os.path.dirname(__file__)
         if template_xml_path is None and template_defaults is None:
-            template_xml_path = os.path.join(file_dir, 'data', 'default_raw_template.xml')
+            template_xml_path = os.path.join(file_dir, 'data',
+                                            'default_raw_template.xml')
 
-        elif template_xml_path is None and template_defaults is not None:
+        elif template_xml_path is None and \
+        template_defaults is not None:
             if template_defaults in default_templates_paths:
                 file_name = default_templates_paths[template_defaults]
             else:
                 print('template_defaults paramter must be in:')
                 for key in default_templates_paths:
                     print(key)
-                raise ValueError("%s is not valid template_defaults parameter" % template_defaults)
-            template_xml_path = os.path.join(file_dir, 'data', file_name)
+                raise ValueError("%s is not valid template_defaults \
+                                 parameter" % template_defaults)
+            template_xml_path = os.path.join(file_dir, 'data',
+                                             file_name)
 
-        elif template_xml_path is not None and template_defaults is None:
+        elif template_xml_path is not None and \
+        template_defaults is None:
             template_xml_path = template_xml_path
         else:
-            raise ValueError('template_xml_path and template_defaults cannot both be specified.')
+            raise ValueError('template_xml_path and template_defaults \
+                              cannot both be specified.')
 
         with open(template_xml_path, 'r') as f:
             root = ET.fromstring(f.read())
@@ -199,7 +228,8 @@ class LogViewer(object):
         formations = root.findall('formation')
         num_tracks = len(tracks)
 
-        self.fig, self.axes = plt.subplots(1, num_tracks + 2, sharey = True)
+        self.fig, self.axes = plt.subplots(1, num_tracks + 2,
+                                           sharey = True)
 
         ### add formation background color ###
         for formation in formations:
@@ -207,12 +237,14 @@ class LogViewer(object):
             if 'name' in formation.attrib:
                 name = formation.attrib['name']
             else:
-                raise ValueError('Formation name required in template %s' % template_xml_path)
+                raise ValueError('Formation name required in \
+                                  template %s' % template_xml_path)
 
             if 'color' in formation.attrib:
                 color = formation.attrib['color']
             else:
-                raise ValueError('Color required for formation %s in template %s' % (name, template_xml_path))
+                raise ValueError('Color required for formation %s \
+                           in template %s' % (name, template_xml_path))
 
             if 'alpha' in formation.attrib:
                 alpha = float(formation.attrib['alpha'])
@@ -221,8 +253,16 @@ class LogViewer(object):
 
             formation_top = self.log.tops[name]
             formation_bottom = self.log.next_formation_depth(name)
-            for ax in axes:
-                ax.axhspan(formation_top, formation_bottom, facecolor = color, alpha = alpha)
+            formation_mid = (formation_top + formation_bottom) / 2.0
+            for ax in self.axes:
+                ax.axhspan(formation_top, formation_bottom,
+                           facecolor = color, alpha = alpha)
+
+            for ax in (self.axes[0], self.axes[-1]):
+                ax.text(0.5, formation_mid, name,
+                        verticalalignment = 'center',
+                        horizontalalignment = 'center',
+                        rotation = 90, fontsize = 16)
 
         numbers = None
         ticks = None
@@ -263,26 +303,34 @@ class LogViewer(object):
                 if 'tick_spacing' in track.attrib:
                     tick_spacing = int(track.attrib['tick_spacing'])
                 else:
-                    raise ValueError('tick_spacing is required for depth track.')
+                    raise ValueError('tick_spacing is required for \
+                                                         depth track.')
 
                 if 'number_spacing' in track.attrib:
-                    number_spacing = int(track.attrib['number_spacing'])
+                    number_spacing =int(track.attrib['number_spacing'])
                 else:
-                    raise ValueError('number_spacing is required for depth track.')
+                    raise ValueError('number_spacing is required for \
+                                                         depth track.')
 
                 if 'line_spacing' in track.attrib:
                     line_spacing = int(track.attrib['line_spacing'])
                 else:
-                    raise ValueError('line_spacing is required for depth track.')
+                    raise ValueError('line_spacing is required for \
+                                                         depth track.')
 
                 font_size = 16
                 if 'font_size' in track.attrib:
                     font_size = float(track.attrib['font_size'])
 
                 max_depth = self.log[0].max()
-                ticks = range(0, int(max_depth) + tick_spacing, tick_spacing)
-                numbers = range(0, int(max_depth) + number_spacing, number_spacing)
-                lines = range(0, int(max_depth) + line_spacing, line_spacing)
+                ticks = range(0, int(max_depth) + tick_spacing,
+                              tick_spacing)
+
+                numbers = range(0, int(max_depth) + number_spacing,
+                                number_spacing)
+
+                lines = range(0, int(max_depth) + line_spacing,
+                              line_spacing)
 
                 for n in numbers:
                     ax.text(0.5, n, str(int(n)),
@@ -293,8 +341,11 @@ class LogViewer(object):
 
             elif 'cumulative' in track.attrib:
 
-                if 'left' not in track.attrib or 'right' not in track.attrib:
-                    raise ValueError('left and right values must be specified in cumulative tracks for template %s.' % template_xml_path)
+                if 'left' not in track.attrib or \
+                'right' not in track.attrib:
+                    raise ValueError('left and right values must be \
+                     specified in cumulative tracks for template %s.' \
+                     % template_xml_path)
 
                 left = float(track.attrib['left'])
                 right = float(track.attrib['right'])
@@ -310,18 +361,24 @@ class LogViewer(object):
                 for c, curve in enumerate(track):
                     ### names ###
                     if 'curve_name' not in curve.attrib:
-                        raise ValueError('Curve Name required in template at %s' % template_xml_path)
+                        raise ValueError('Curve Name required in \
+                                   template at %s' % template_xml_path)
                     curve_name = curve.attrib['curve_name']
 
                     if curve_name not in self.log.keys():
-                        raise ValueError('Curve %s not found in log.' % curve_name)
+                        raise ValueError('Curve %s not found in log.' \
+                                         % curve_name)
 
                     if 'fill_color' not in curve.attrib:
-                        raise ValueError('Curve fill_color must be specificied for cumulative track in template at %s' % template_xml_path)
+                        raise ValueError('Curve fill_color must be \
+                                         specificied for cumulative \
+                                         track in template at %s' % \
+                                         template_xml_path)
 
                     fill_color = curve.attrib['fill_color']
 
-                    increase = summation + np.asarray(self.log[curve_name])
+                    increase=summation+np.asarray(self.log[curve_name])
+
                     ax.fill_betweenx(self.log[0],
                                      summation,
                                      increase,
@@ -331,9 +388,10 @@ class LogViewer(object):
 
                     ### display ###
                     if invert_axis:
-                        x = (len(track) - float(c)) / len(track) - 1.0 / (2 * len(track))
+                        x=(len(track) - float(c)) / len(track) - 1.0/ \
+                          (2 * len(track))
                     else:
-                        x = float(c) / len(track) + 1.0 / (2 * len(track))
+                        x=float(c) / len(track) + 1.0 /(2 * len(track))
 
                     ax.text(x, 1.03, curve_name, rotation = 90,
                             horizontalalignment = 'center',
@@ -343,11 +401,12 @@ class LogViewer(object):
                             color = fill_color)
 
                 if 'major_lines' in track.attrib:
-                    num_major_lines = int(track.attrib['major_lines']) + 1
+                    num_major_lines=int(track.attrib['major_lines']) +1
                     dist = abs(left - right) / num_major_lines
                     major_lines = np.arange(left + dist, right, dist)
                     for m in major_lines:
-                        ax.plot((m, m), (0, self.log[0].max()), color = '#c0c0c0', lw = 0.5)
+                        ax.plot((m, m), (0, self.log[0].max()),
+                                color = '#c0c0c0', lw = 0.5)
 
                 ax.set_xlim(left, right)
                 if invert_axis:
@@ -361,10 +420,12 @@ class LogViewer(object):
                     if 'curve_name' in curve.attrib:
                         curve_name = curve.attrib['curve_name']
                     else:
-                        raise ValueError('Curve Name required in template at %s' % template_xml_path)
+                        raise ValueError('Curve Name required in \
+                                   template at %s' % template_xml_path)
 
                     if curve_name not in self.log.keys():
-                        raise ValueError('Curve %s not found in log.' % curve_name)
+                        raise ValueError('Curve %s not found in log.' \
+                                         % curve_name)
 
                     ### style and scale ###
 
@@ -378,13 +439,16 @@ class LogViewer(object):
                         right_label = curve.attrib['right'] + ' '
 
                     if left is None:
-                        print('Adjust template at %s.' % template_xml_path)
-                        raise ValueError('Left X axis not found for curve %s.'\
-                                         % curve_name)
+                        print('Adjust template at %s.' % \
+                              template_xml_path)
+                        raise ValueError('Left X axis not found for \
+                                          curve %s.' % curve_name)
                     if right is None:
-                        print('Adjust template at %s.' % template_xml_path)
-                        raise ValueError('Right X axis not found for curve %s.'\
-                                         % curve_name)
+                        print('Adjust template at %s.' \
+                              % template_xml_path)
+
+                        raise ValueError('Right X axis not found for \
+                                         curve %s.' % curve_name)
 
                     line_style = '-'
                     color = '#000000'
@@ -404,7 +468,7 @@ class LogViewer(object):
                     if 'marker' in curve.attrib:
                         marker = curve.attrib['marker']
                     if 'marker_size' in curve.attrib:
-                        marker_size = float(curve.attrib['marker_size'])
+                        marker_size =float(curve.attrib['marker_size'])
 
 
                     if scale == 'log':
@@ -412,6 +476,19 @@ class LogViewer(object):
                         x = self.log[curve_name]
                         m = None
                         b = None
+
+                        if 'left_color_value' in curve.attrib:
+                            left_color_value = \
+                                float(curve.attrib['left_color_value'])
+                        else:
+                            left_color_value = left
+
+                        if 'right_color_value' in curve.attrib:
+                            right_color_value = \
+                               float(curve.attrib['right_color_value'])
+                        else:
+                            right_color_value = right
+
                     else:
                         ax.set_xlim(0,1)
                         m = (1 - 0) / (right - left)
@@ -420,9 +497,25 @@ class LogViewer(object):
                         left = 0
                         right = 1
 
+                        if 'left_color_value' in curve.attrib:
+                            left_color_value = \
+                                float(curve.attrib['left_color_value'])
+                            left_color_value = m * left_color_value + b
+                        else:
+                            left_color_value = left
+
+                        if 'right_color_value' in curve.attrib:
+                            right_color_value = \
+                               float(curve.attrib['right_color_value'])
+                            right_color_value=m * right_color_value + b
+                        else:
+                            right_color_value = right
+
                     ### label ###
                     if 'display_name' in curve.attrib:
-                        self._display_name_to_curve_name[curve.attrib['display_name']] = curve_name
+                        self._display_name_to_curve_name\
+                            [curve.attrib['display_name']] = curve_name
+
                         ax.text(0.5, 0.98 + 0.035 * (c + 1),
                                 curve.attrib['display_name'],
                                 horizontalalignment = 'center',
@@ -454,25 +547,58 @@ class LogViewer(object):
                         if 'fill_color' in curve.attrib:
                             fill_color = curve.attrib['fill_color']
 
-                        if curve.attrib['fill'] == 'left':
+                            if curve.attrib['fill'] == 'left':
+                                baseline = left
+                            elif curve.attrib['fill'] == 'right':
+                                baseline = right
+
                             ax.fill_betweenx(self.log[0],
-                                             left,
+                                             baseline,
                                              x,
                                              color = fill_color)
 
-                        elif curve.attrib['fill'] == 'right':
-                             ax.fill_betweenx(self.log[0],
-                                              x,
-                                              right,
-                                              color = fill_color)
+                        elif 'fill_color_map' in curve.attrib:
+                            cmap_name = curve.attrib['fill_color_map']
+
+                            span = \
+                              abs(left_color_value - right_color_value)
+
+                            cmap = plt.get_cmap(cmap_name)
+
+                            if len(np.unique(x)) < 50:
+                                colored_index = np.unique(x)
+                            else:
+                                color_index=np.arange(left_color_value,\
+                                                     right_color_value,\
+                                                     span / 50.0)
+
+                            if curve.attrib['fill'] == 'left':
+                                baseline = np.ones(len(x)) * \
+                                           min(left_color_value, left)
+
+                            elif curve.attrib['fill'] == 'right':
+                                baseline = np.ones(len(x)) * \
+                                          max(right_color_value, right)
+
+                            for ci in sorted(color_index):
+                                ci_value = (ci - left_color_value)/span
+                                color = cmap(ci_value)
+                                ax.fill_betweenx(self.log[0],
+                                                 baseline,
+                                                 x,
+                                                 where = x >= ci,
+                                                 color = color)
 
                     if 'right_cutoff_fill' in curve.attrib:
 
                         fill_color = '#000000'
                         if 'right_cutoff_fill_color' in curve.attrib:
-                            fill_color = curve.attrib['right_cutoff_fill_color']
+                            fill_color = \
+                                curve.attrib['right_cutoff_fill_color']
 
-                        cutoff_value = float(curve.attrib['right_cutoff_fill'])
+                        cutoff_value = \
+                               float(curve.attrib['right_cutoff_fill'])
+
                         if scale != 'log':
                             v = m * cutoff_value + b
                         else:
@@ -483,15 +609,19 @@ class LogViewer(object):
                                          color = fill_color,
                                          where = v < x)
                         ax.plot(v * np.ones(len(self.log[0][v < x])),
-                                self.log[0][v < x], c = "#000000", lw = 0.5)
+                                self.log[0][v < x], c = "#000000",
+                                lw = 0.5)
 
                     if 'left_cutoff_fill' in curve.attrib:
 
                         fill_color = '#000000'
                         if 'left_cutoff_fill_color' in curve.attrib:
-                            fill_color = curve.attrib['left_cutoff_fill_color']
+                            fill_color = \
+                                 curve.attrib['left_cutoff_fill_color']
 
-                        cutoff_value = float(curve.attrib['left_cutoff_fill'])
+                        cutoff_value = \
+                                float(curve.attrib['left_cutoff_fill'])
+
                         if scale != 'log':
                             v = m * cutoff_value + b
                         else:
@@ -501,29 +631,39 @@ class LogViewer(object):
                                          x,
                                          color = fill_color,
                                          where = x < v)
+
                         ax.plot(v * np.ones(len(self.log[0][x < v])),
-                                self.log[0][x < v], c = "#000000", lw = 0.5)
+                                self.log[0][x < v], c = "#000000",
+                                lw = 0.5)
 
                     if 'left_crossover' in curve.attrib:
 
                         fill_color = '#000000'
                         if 'left_crossover_fill_color' in curve.attrib:
-                            fill_color = curve.attrib['left_crossover_fill_color']
+                            fill_color = \
+                              curve.attrib['left_crossover_fill_color']
 
                         left_curve = curve.attrib['left_crossover']
                         if left_curve not in self.log.keys():
-                            raise ValueError('Curve %s not found in log.' % left_curve)
+                            raise ValueError('Curve %s not found in \
+                                             log.' % left_curve)
 
-                        if 'left_crossover_left' not in curve.attrib and \
-                           'left_crossover_right' not in curve.attrib:
-                           raise ValueError('left and right crossover values not found in template %s' \
-                                             % template_xml_path)
+                        if 'left_crossover_left' not in curve.attrib \
+                        and 'left_crossover_right' not in curve.attrib:
+                           raise ValueError('left and right crossover \
+                                     values not found in template %s' \
+                                     % template_xml_path)
 
-                        left_crossover_left = float(curve.attrib['left_crossover_left'])
-                        left_crossover_right = float(curve.attrib['left_crossover_right'])
+                        left_crossover_left = \
+                             float(curve.attrib['left_crossover_left'])
+
+                        left_crossover_right = \
+                            float(curve.attrib['left_crossover_right'])
 
                         if scale != 'log':
-                            m = (1 - 0) / (left_crossover_right - left_crossover_left)
+                            m = (1 - 0) / (left_crossover_right - \
+                                                   left_crossover_left)
+
                             b = -m * left_crossover_left
                             v = m * self.log[left_curve] + b
                         else:
@@ -539,22 +679,29 @@ class LogViewer(object):
 
                         fill_color = '#000000'
                         if 'right_crossover_fill_color' in curve.attrib:
-                            fill_color = curve.attrib['right_crossover_fill_color']
+                            fill_color = \
+                             curve.attrib['right_crossover_fill_color']
 
                         left_curve = curve.attrib['right_crossover']
                         if left_curve not in self.log.keys():
-                            raise ValueError('Curve %s not found in log.' % left_curve)
+                            raise ValueError('Curve %s not found in \
+                                                    log.' % left_curve)
 
-                        if 'right_crossover_left' not in curve.attrib and \
-                           'right_crossover_right' not in curve.attrib:
-                           raise ValueError('left and right crossover values not found in template %s' \
-                                            % template_xml_path)
+                        if 'right_crossover_left' not in curve.attrib \
+                        and 'right_crossover_right' not in curve.attrib:
+                           raise ValueError('left and right crossover \
+                                     values not found in template %s' \
+                                     % template_xml_path)
 
-                        left_crossover_left = float(curve.attrib['right_crossover_left'])
-                        left_crossover_right = float(curve.attrib['right_crossover_right'])
+                        left_crossover_left = \
+                            float(curve.attrib['right_crossover_left'])
+
+                        left_crossover_right = \
+                           float(curve.attrib['right_crossover_right'])
 
                         if scale != 'log':
-                            m = (1 - 0) / (left_crossover_right - left_crossover_left)
+                            m = (1 - 0) / (left_crossover_right - \
+                                                   left_crossover_left)
                             b = -m * left_crossover_left
                             v = m * self.log[left_curve] + b
                         else:
@@ -574,17 +721,21 @@ class LogViewer(object):
                                  marker = marker,
                                  ms = marker_size)[0]
 
-                    self._edit_curve_lines[curve_name] = (curve_line, m, b)
+                    self._edit_curve_lines[curve_name] = \
+                                                     (curve_line, m, b)
 
                 if scale == 'log':
-                    ax.xaxis.grid(True, which = 'both', color = '#e0e0e0')
+                    ax.xaxis.grid(True, which = 'both',color='#e0e0e0')
 
                 elif 'major_lines' in track.attrib:
-                    num_major_lines = int(track.attrib['major_lines']) + 1
+                    num_major_lines = \
+                                   int(track.attrib['major_lines']) + 1
+
                     dist = abs(left - right) / num_major_lines
                     major_lines = np.arange(left + dist, right, dist)
                     for m in major_lines:
-                        ax.plot((m, m), (0, self.log[0].max()), color = '#c0c0c0', lw = 0.5)
+                        ax.plot((m, m), (0, self.log[0].max()),
+                                color = '#c0c0c0', lw = 0.5)
 
         if max_track_curves < 5:
             max_track_curves = 5
@@ -597,16 +748,19 @@ class LogViewer(object):
         track_widths = track_widths / np.sum(track_widths)
         track_locations = [0]
         for t in range(1, len(track_widths)):
-            track_locations.append(track_locations[t - 1] + track_widths[t - 1])
+            track_locations.append(track_locations[t - 1] + \
+                                                   track_widths[t - 1])
 
         for a, ax in enumerate(self.axes):
             post = ax.get_position()
-            new_post = (track_locations[a], 0.01, track_widths[a], post.height)
+            new_post = (track_locations[a], 0.01, track_widths[a],
+                        post.height)
             ax.set_position(new_post)
             if ticks is not None:
                 ax.set_yticks(lines, minor = False)
                 ax.set_yticks(ticks, minor = True)
-                ax.tick_params(axis = 'y', direction = 'inout', length = 6, width = 1,
+                ax.tick_params(axis = 'y', direction = 'inout',
+                               length = 6, width = 1,
                                colors = '#000000', which = 'minor')
                 if a not in depth_track_numbers:
                     ax.yaxis.grid(True, which = 'major')
@@ -619,7 +773,7 @@ class LogViewer(object):
                 if track_title is not None:
                     for i in range(max_track_curves):
                         track_title += '\n'
-                    ax.set_title(track_title, fontweight='bold')
+                    ax.set_title(track_title, fontweight = 'bold')
 
         if top is not None and bottom is not None:
             plt.ylim((top, bottom))
@@ -629,9 +783,11 @@ class LogViewer(object):
 
     def show(self, edit_mode = False, window_location = (10, 30)):
         """
-        Calls matplotlib.pyplot.show() to display log viewer. If edit_mode == True, it includes options
-        to graphically edit curve data, and stores these changes within the LogViewer object. After
-        editing is finished, access the updated data with the .log property.
+        Calls matplotlib.pyplot.show() to display log viewer. If
+        edit_mode == True, it includes options to graphically edit
+        curve data, and stores these changes within the LogViewer
+        object. After editing is finished, access the updated
+        :class:`petropy.Log` data with the .log property.
 
         Parameters
         -----------
@@ -643,7 +799,7 @@ class LogViewer(object):
             Second value is pixels from the top of the screen.
 
         Example
-        --------
+        -------
         >>> import petropy as ptr
         >>> log = ptr.log_data('WFMP') # sample Wolfcamp log
         >>> viewer = ptr.LogViewer(log) # default triple-combo template
@@ -672,13 +828,18 @@ class LogViewer(object):
             self.edit_fig, self.edit_axes = plt.subplots(1)
 
             rax = plt.axes([0, 0, 1, 1])
-            self._radio_button = RadioButtons(rax, ('No Edit', 'Manual Edit', 'Bulk Shift'))
+            self._radio_button = RadioButtons(rax, ('No Edit',
+                                          'Manual Edit', 'Bulk Shift'))
             self._radio_button.on_clicked(self._radio_click)
 
-            self.fig.canvas.mpl_connect('pick_event', self._curve_pick)
-            self.fig.canvas.mpl_connect('button_press_event', self._edit_lock_toggle)
-            self.fig.canvas.mpl_connect('button_release_event', self._edit_lock_toggle)
-            self.fig.canvas.mpl_connect('motion_notify_event', self._draw_curve)
+            self.fig.canvas.mpl_connect('pick_event',
+                                        self._curve_pick)
+            self.fig.canvas.mpl_connect('button_press_event',
+                                        self._edit_lock_toggle)
+            self.fig.canvas.mpl_connect('button_release_event',
+                                        self._edit_lock_toggle)
+            self.fig.canvas.mpl_connect('motion_notify_event',
+                                        self._draw_curve)
 
             edit_x = x + dx + 17
             edit_y = y
@@ -691,61 +852,78 @@ class LogViewer(object):
 
     def _curve_pick(self, event):
         """
-        Event handler for selecting a curve to edit. Results in self._edit_curve being set to matplotlib
-        text object. Connected on line 610 with 'pick_event'.
+        Event handler for selecting a curve to edit. Results in
+        self._edit_curve being set to matplotlib text object. Connected
+        on line 765 with 'pick_event'.
         """
         if self._edit_curve is not None:
-            self._edit_curve.set_bbox({'facecolor':'white', 'edgecolor': 'white', 'alpha': 0})
+            self._edit_curve.set_bbox({'facecolor':'white',
+                                       'edgecolor': 'white',
+                                       'alpha': 0})
 
         self._edit_curve = event.artist
 
         if self._radio_button.value_selected != 'No Edit':
-            self._edit_curve.set_bbox({'facecolor':'lightgoldenrodyellow', 'edgecolor': 'lightgoldenrodyellow', 'alpha': 1})
+            self._edit_curve.set_bbox({'facecolor':'khaki',
+                                       'edgecolor': 'khaki',
+                                       'alpha': 1})
 
         self.fig.canvas.draw()
 
     def _radio_click(self, label):
         """
-        Event handler for selecting which editing type to use. Connected on line 608 with on_click method
-        for RadioButtons object.
+        Event handler for selecting which editing type to use.
+        Connected on line 763 with on_click method for RadioButtons
+        object.
         """
         if label == 'No Edit':
             if self._edit_curve is not None:
-                self._edit_curve.set_bbox({'facecolor':'white', 'edgecolor': 'white', 'alpha': 0})
+                self._edit_curve.set_bbox({'facecolor':'white',
+                                           'edgecolor': 'white',
+                                           'alpha': 0})
         elif label == 'Manual Edit':
             if self._edit_curve is not None:
-                self._edit_curve.set_bbox({'facecolor':'lightgoldenrodyellow', 'edgecolor': 'lightgoldenrodyellow', 'alpha': 1})
+                self._edit_curve.set_bbox({'facecolor':'khaki',
+                                           'edgecolor': 'khaki',
+                                           'alpha': 1})
         elif label == 'Bulk Shift':
             if self._edit_curve is not None:
-                self._edit_curve.set_bbox({'facecolor':'lightgoldenrodyellow', 'edgecolor': 'lightgoldenrodyellow', 'alpha': 1})
+                self._edit_curve.set_bbox({'facecolor':'khaki',
+                                           'edgecolor': 'khaki',
+                                           'alpha': 1})
 
         self.fig.canvas.draw()
 
     def _edit_lock_toggle(self, event):
         """
-        Event handler to check for correct axis associated with selected curve. Will allow _draw_curve to
-        function if click is in proper axis based on the self._edit_curve property set with _curve_pick.
-        Connected on lines 611 and 612 to 'button_press_event' and 'button_release_event'.
+        Event handler to check for correct axis associated with
+        selected curve. Will allow _draw_curve to function if click is
+        in proper axis based on the self._edit_curve property set with
+        _curve_pick. Connected on lines 767 and 769 to
+        :code:`button_press_event` and :code:`button_release_event`.
         """
 
         if self._edit_curve and hasattr(event, 'inaxes'):
             if event.inaxes:
                 ax_num = np.where(self.axes == event.inaxes)[0]
                 if len(ax_num) > 0:
-                    curve_num = self.axes.tolist().index(self._edit_curve.axes)
+                    curve_num = \
+                        self.axes.tolist().index(self._edit_curve.axes)
                     if ax_num == curve_num:
                         self._edit_lock = not self._edit_lock
 
     def _draw_curve(self, event):
         """
-        Event handler for changing data in the figure and in the log object. Connected on line
-        613 with 'motion_notify_event'.
+        Event handler for changing data in the figure and in the log
+        object. Connected on line 771 with :code:`motion_notify_event`.
         """
 
-        if self._radio_button.value_selected == 'Manual Edit' and self._edit_lock:
+        if self._radio_button.value_selected == 'Manual Edit' and \
+        self._edit_lock:
             x, y = event.xdata, event.ydata
 
-            curve_name = self._display_name_to_curve_name[self._edit_curve.get_text()]
+            curve_name = \
+          self._display_name_to_curve_name[self._edit_curve.get_text()]
 
             cursor_depth_index = np.argmin(np.abs(self.log[0] - y))
             depth = self.log[0][cursor_depth_index]
@@ -762,11 +940,13 @@ class LogViewer(object):
 
             self.fig.canvas.draw()
 
-        elif self._radio_button.value_selected == 'Bulk Shift' and self._edit_lock:
+        elif self._radio_button.value_selected == 'Bulk Shift' and \
+        self._edit_lock:
 
             x, y = event.xdata, event.ydata
 
-            curve_name = self._display_name_to_curve_name[self._edit_curve.get_text()]
+            curve_name = \
+          self._display_name_to_curve_name[self._edit_curve.get_text()]
 
             cursor_depth_index = np.argmin(np.abs(self.log[0] - y))
             depth = self.log[0][cursor_depth_index]